Removed files not needed anymore
