+++ src/Developurr.Orderly.Application/Query/Customer/GetCustomer/GetCustomerInput.cs
@@ -1,0 +1,3 @@
+namespace Developurr.Orderly.Application.Query.Customer.GetCustomer;
+
+public record GetCustomerInput(string CustomerId);
