+++ src/Developurr.Orderly.Application/Query/Customer/GetCustomer/GetCustomerUseCase.cs
@@ -1,9 +1,8 @@
-using Developurr.Orderly.Application.Command;
 using Developurr.Orderly.Domain.Customer;
 
-namespace Developurr.Orderly.Application.UseCase.Customer.GetCustomer;
+namespace Developurr.Orderly.Application.Query.Customer.GetCustomer;
 
-public class GetCustomerUseCase : IUseCase<GetCustomerInput, GetCustomerOutput>
+public sealed class GetCustomerUseCase : IQuery<GetCustomerInput, GetCustomerOutput>
 {
     private readonly ICustomerRepository _customerRepository;
 
@@ -17,18 +16,23 @@ public class GetCustomerUseCase : IUseCase<GetCustomerInput, GetCustomerOutput>
         CancellationToken cancellationToken
     )
     {
-        var customer = await _customerRepository.GetByIdAsync(input.CustomerId, cancellationToken);
+        var customer = await _customerRepository.GetByIdAsync(
+            input.CustomerId,
+            cancellationToken
+        );
 
         return new GetCustomerOutput(
+            customer.Id.Format(),
+            customer.SalesConsultant.Format(),
             customer.Cnpj.Format(),
             customer.CorporateName,
             customer.TaxId,
             customer.TradeName,
             customer.Segment,
-            customer.BillingEmail?.Format() ?? "",
+            customer.BillingEmail.Format(),
             customer.NfeEmail.Format(),
-            customer.Landline?.Value ?? "",
-            customer.Mobile?.Value ?? "",
+            customer.Landline.Value,
+            customer.Mobile.Value,
             customer.Observation
         );
     }
