--- src/Developurr.Orderly.Application/Query/Customer/GetCustomerInput.cs
@@ -1,3 +1,0 @@
-namespace Developurr.Orderly.Application.Query.Customer;
-
-public record GetCustomerInput(string CustomerId);
