+++ src/Developurr.Orderly.Application/Query/Order/GetOrder/GetOrderInput.cs
@@ -1,0 +1,3 @@
+namespace Developurr.Orderly.Application.Query.Order.GetOrder;
+
+public record GetOrderInput(string OrderId);
