+++ src/Developurr.Orderly.Application/Query/Order/GetOrder/GetOrderOutput.cs
@@ -1,0 +1,5 @@
+namespace Developurr.Orderly.Application.Query.Order.GetOrder;
+
+public class GetOrderOutput(
+    string Price
+);
