--- src/Developurr.Orderly.Application/Query/Order/GetOrderInput.cs
@@ -1,3 +1,0 @@
-namespace Developurr.Orderly.Application.Query.Order;
-
-public record GetOrderInput(string OrderId);
