--- src/Developurr.Orderly.Application/Query/Order/GetOrderOutput.cs
@@ -1,5 +1,0 @@
-namespace Developurr.Orderly.Application.Query.Order;
-
-public class GetOrderOutput(
-    string Price
-);
