--- src/Developurr.Orderly.Application/UseCase/Customer/CreateCustomer/CreateCustomerOutput.cs
@@ -1,3 +1,0 @@
-namespace Developurr.Orderly.Application.UseCase.Customer.CreateCustomer;
-
-public record CreateCustomerOutput(string CustomerId);
