--- src/Developurr.Orderly.Application/UseCase/Order/CreateOrder/CreateOrderInput.cs
@@ -1,8 +1,0 @@
-namespace Developurr.Orderly.Application.UseCase.Order.CreateOrder;
-
-public record CreateOrderInput(
-    string CustomerId,
-    string SalesConsultantId,
-    string ShippingId,
-    string Price
-);
