--- src/Developurr.Orderly.Application/UseCase/Order/CreateOrder/CreateOrderOutput.cs
@@ -1,3 +1,0 @@
-namespace Developurr.Orderly.Application.UseCase.Order.CreateOrder;
-
-public record CreateOrderOutput(string OrderId);
