--- src/Developurr.Orderly.Application/UseCase/Order/DeleteOrder/DeleteOrderInput.cs
@@ -1,3 +1,0 @@
-namespace Developurr.Orderly.Application.UseCase.Order.DeleteOrder;
-
-public record DeleteOrderInput(string OrderId);
