--- src/Developurr.Orderly.Application/UseCase/Order/DeleteOrder/DeleteOrderOutput.cs
@@ -1,3 +1,0 @@
-namespace Developurr.Orderly.Application.UseCase.Order.DeleteOrder;
-
-public record DeleteOrderOutput(string OrderId);
