--- src/Developurr.Orderly.Application/UseCase/Order/GetOrder/GetOrderInput.cs
@@ -1,3 +1,0 @@
-namespace Developurr.Orderly.Application.UseCase.Order.GetOrder;
-
-public record GetOrderInput(string OrderId);
