--- src/Developurr.Orderly.Application/UseCase/Order/GetOrder/GetOrderOutput.cs
@@ -1,5 +1,0 @@
-namespace Developurr.Orderly.Application.UseCase.Order.GetOrder;
-
-public class GetOrderOutput(
-    string Price
-);
